feat(core): expose `async_runtime` module
