+++ .changes/expose-async-runtime.md
@@ -1,0 +1,5 @@
+---
+"tauri": patch
+---
+
+Expose `async_runtime` module.
