+++ core/tauri/src/api/mod.rs
@@ -68,33 +68,6 @@ pub struct PackageInfo {
 // Not public API
 #[doc(hidden)]
 pub mod private {
-  // Core API only.
-  pub mod async_runtime {
-    use once_cell::sync::OnceCell;
-    use tokio::runtime::Runtime;
-    pub use tokio::sync::{
-      mpsc::{channel, Receiver, Sender},
-      Mutex, RwLock,
-    };
-
-    use std::future::Future;
-
-    static RUNTIME: OnceCell<Runtime> = OnceCell::new();
-
-    pub fn block_on<F: Future>(task: F) -> F::Output {
-      let runtime = RUNTIME.get_or_init(|| Runtime::new().unwrap());
-      runtime.block_on(task)
-    }
-
-    pub fn spawn<F>(task: F)
-    where
-      F: Future + Send + 'static,
-      F::Output: Send + 'static,
-    {
-      let runtime = RUNTIME.get_or_init(|| Runtime::new().unwrap());
-      runtime.spawn(task);
-    }
-  }
   pub use once_cell::sync::OnceCell;
 
   pub trait AsTauriContext {
